+++ resources/js/actions/App/Http/Controllers/UserRoleController.ts
@@ -270,7 +270,7 @@ store.form = storeForm
 * @see app/Http/Controllers/UserRoleController.php:98
 * @route '/users/{user}/roles/{role}'
 */
-export const destroy = (args: { user: number | { id: number }, role: string | number | { id: string | number } } | [user: number | { id: number }, role: string | number | { id: string | number } ], options?: RouteQueryOptions): RouteDefinition<'delete'> => ({
+export const destroy = (args: { user: number | { id: number }, role: number | { id: number } } | [user: number | { id: number }, role: number | { id: number } ], options?: RouteQueryOptions): RouteDefinition<'delete'> => ({
     url: destroy.url(args, options),
     method: 'delete',
 })
@@ -285,7 +285,7 @@ destroy.definition = {
 * @see app/Http/Controllers/UserRoleController.php:98
 * @route '/users/{user}/roles/{role}'
 */
-destroy.url = (args: { user: number | { id: number }, role: string | number | { id: string | number } } | [user: number | { id: number }, role: string | number | { id: string | number } ], options?: RouteQueryOptions) => {
+destroy.url = (args: { user: number | { id: number }, role: number | { id: number } } | [user: number | { id: number }, role: number | { id: number } ], options?: RouteQueryOptions) => {
     if (Array.isArray(args)) {
         args = {
             user: args[0],
@@ -315,7 +315,7 @@ destroy.url = (args: { user: number | { id: number }, role: string | number | { 
 * @see app/Http/Controllers/UserRoleController.php:98
 * @route '/users/{user}/roles/{role}'
 */
-destroy.delete = (args: { user: number | { id: number }, role: string | number | { id: string | number } } | [user: number | { id: number }, role: string | number | { id: string | number } ], options?: RouteQueryOptions): RouteDefinition<'delete'> => ({
+destroy.delete = (args: { user: number | { id: number }, role: number | { id: number } } | [user: number | { id: number }, role: number | { id: number } ], options?: RouteQueryOptions): RouteDefinition<'delete'> => ({
     url: destroy.url(args, options),
     method: 'delete',
 })
@@ -325,7 +325,7 @@ destroy.delete = (args: { user: number | { id: number }, role: string | number |
 * @see app/Http/Controllers/UserRoleController.php:98
 * @route '/users/{user}/roles/{role}'
 */
-const destroyForm = (args: { user: number | { id: number }, role: string | number | { id: string | number } } | [user: number | { id: number }, role: string | number | { id: string | number } ], options?: RouteQueryOptions): RouteFormDefinition<'post'> => ({
+const destroyForm = (args: { user: number | { id: number }, role: number | { id: number } } | [user: number | { id: number }, role: number | { id: number } ], options?: RouteQueryOptions): RouteFormDefinition<'post'> => ({
     action: destroy.url(args, {
         [options?.mergeQuery ? 'mergeQuery' : 'query']: {
             _method: 'DELETE',
@@ -340,7 +340,7 @@ const destroyForm = (args: { user: number | { id: number }, role: string | numbe
 * @see app/Http/Controllers/UserRoleController.php:98
 * @route '/users/{user}/roles/{role}'
 */
-destroyForm.delete = (args: { user: number | { id: number }, role: string | number | { id: string | number } } | [user: number | { id: number }, role: string | number | { id: string | number } ], options?: RouteQueryOptions): RouteFormDefinition<'post'> => ({
+destroyForm.delete = (args: { user: number | { id: number }, role: number | { id: number } } | [user: number | { id: number }, role: number | { id: number } ], options?: RouteQueryOptions): RouteFormDefinition<'post'> => ({
     action: destroy.url(args, {
         [options?.mergeQuery ? 'mergeQuery' : 'query']: {
             _method: 'DELETE',